Makke file initDB, Ñ, add filed dateTime in class Restaurant, correct imports

+++ src/main/java/ru/grigoriev/graduationproject/model/Menu.java
@@ -1,15 +1,15 @@
 package ru.grigoriev.graduationproject.model;
 
-import lombok.*;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 import org.hibernate.annotations.OnDelete;
 import org.hibernate.annotations.OnDeleteAction;
-import org.hibernate.validator.constraints.Range;
 
 import javax.persistence.*;
 import javax.validation.constraints.Min;
 import javax.validation.constraints.NotNull;
-import java.awt.*;
-import java.time.LocalDate;
 import java.time.LocalDateTime;
 
 @Entity
@@ -18,11 +18,11 @@ import java.time.LocalDateTime;
 @Getter
 @Setter
 @Table(name = "menus", uniqueConstraints = {@UniqueConstraint(columnNames = {"restaurant_id", "dish_id", "date_time"}, name = "menu_unique_restaurant_dish_datetime_idx")})
-public class Menu extends AbstractBaseEntity  {
+public class Menu extends AbstractBaseEntity {
 
     @Column(name = "price", nullable = false)
     @Min(value = 1, message = "Price should be over 1")
-    private int price;
+    private double price;
 
     @JoinColumn(name = "dish_id", nullable = false)
     @ManyToOne(fetch = FetchType.EAGER)
